// src/entities/product/model/productSchema.ts
import { z } from "zod";
import { SIZES, COLLECTIONS, TAGS, MOCK_COLORS } from "@/shared/lib/constants";

export const rawProductSchema = z.object({
  id: z.number(),
  title: z.string(),
  price: z.number(),
  description: z.string(),
  category: z.string(),
  image: z.string(),
  isDiscount: z.boolean().optional(),
});

export const productSchema = rawProductSchema.extend({
  rating: z.object({
    rate: z.number(),
    count: z.number(),
  }),
  sizes: z.array(z.enum(SIZES)),
  brand: z.string(),
  colors: z.array(z.enum(MOCK_COLORS)),
  collection: z.array(z.enum(COLLECTIONS)),
  tags: z.array(z.enum(TAGS)),
  isDiscount: z.boolean(),
  salePercent: z.number(),
  isSoldOut: z.boolean(),
  isAlmostSoldOut: z.boolean(),
});

export type RawProduct = z.infer<typeof rawProductSchema>;
export type Product = z.infer<typeof productSchema>;